engine: Move more sst functions into engine_rocks
