+++ components/engine/src/rocks/util/mod.rs
@@ -22,35 +22,16 @@ use self::engine_metrics::{
     ROCKSDB_TOTAL_SST_FILES_SIZE,
 };
 use crate::rocks::load_latest_options;
-use crate::rocks::set_external_sst_file_global_seq_no;
 use crate::rocks::supported_compression;
 use crate::rocks::{
     CColumnFamilyDescriptor, ColumnFamilyOptions, CompactOptions, CompactionOptions,
     DBCompressionType, DBOptions, Env, Range, SliceTransform, DB,
 };
 use crate::{Error, Result, ALL_CFS, CF_DEFAULT};
-use tikv_util::file::calc_crc32;
 
 pub use self::event_listener::EventListener;
 pub use self::metrics_flusher::MetricsFlusher;
 pub use crate::rocks::CFHandle;
-
-/// Copies the source file to a newly created file.
-pub fn copy_and_sync<P: AsRef<Path>, Q: AsRef<Path>>(from: P, to: Q) -> io::Result<u64> {
-    if !from.as_ref().is_file() {
-        return Err(io::Error::new(
-            ErrorKind::InvalidInput,
-            "the source path is not an existing regular file",
-        ));
-    }
-
-    let mut reader = File::open(from)?;
-    let mut writer = File::create(to)?;
-
-    let res = io::copy(&mut reader, &mut writer)?;
-    writer.sync_all()?;
-    Ok(res)
-}
 
 // Zlib and bzip2 are too slow.
 const COMPRESSION_PRIORITY: [DBCompressionType; 3] = [
@@ -514,95 +495,6 @@ pub fn compact_files_in_range_cf(
     opts.set_max_subcompactions(max_subcompactions as i32);
     opts.set_output_file_size_limit(output_file_size_limit);
     db.compact_files_cf(cf, &opts, &input_files, output_level)?;
-
-    Ok(())
-}
-
-/// Prepares the SST file for ingestion.
-/// The purpose is to make the ingestion retryable when using the `move_files` option.
-/// Things we need to consider here:
-/// 1. We need to access the original file on retry, so we should make a clone
-///    before ingestion.
-/// 2. `RocksDB` will modified the global seqno of the ingested file, so we need
-///    to modified the global seqno back to 0 so that we can pass the checksum
-///    validation.
-/// 3. If the file has been ingested to `RocksDB`, we should not modified the
-///    global seqno directly, because that may corrupt RocksDB's data.
-#[cfg(target_os = "linux")]
-pub fn prepare_sst_for_ingestion<P: AsRef<Path>, Q: AsRef<Path>>(path: P, clone: Q) -> Result<()> {
-    use std::os::linux::fs::MetadataExt;
-
-    let path = path.as_ref().to_str().unwrap();
-    let clone = clone.as_ref().to_str().unwrap();
-
-    if Path::new(clone).exists() {
-        fs::remove_file(clone).map_err(|e| format!("remove {}: {:?}", clone, e))?;
-    }
-
-    let meta = fs::metadata(path).map_err(|e| format!("read metadata from {}: {:?}", path, e))?;
-
-    if meta.st_nlink() == 1 {
-        // RocksDB must not have this file, we can make a hard link.
-        fs::hard_link(path, clone)
-            .map_err(|e| format!("link from {} to {}: {:?}", path, clone, e))?;
-    } else {
-        // RocksDB may have this file, we should make a copy.
-        copy_and_sync(path, clone)
-            .map_err(|e| format!("copy from {} to {}: {:?}", path, clone, e))?;
-    }
-
-    Ok(())
-}
-
-#[cfg(not(target_os = "linux"))]
-pub fn prepare_sst_for_ingestion<P: AsRef<Path>, Q: AsRef<Path>>(path: P, clone: Q) -> Result<()> {
-    let path = path.as_ref().to_str().unwrap();
-    let clone = clone.as_ref().to_str().unwrap();
-    if !Path::new(clone).exists() {
-        copy_and_sync(path, clone)
-            .map_err(|e| format!("copy from {} to {}: {:?}", path, clone, e))?;
-    }
-    Ok(())
-}
-
-pub fn validate_sst_for_ingestion<P: AsRef<Path>>(
-    db: &DB,
-    cf: &str,
-    path: P,
-    expected_size: u64,
-    expected_checksum: u32,
-) -> Result<()> {
-    let path = path.as_ref().to_str().unwrap();
-    let f = File::open(path)?;
-
-    let meta = f.metadata()?;
-    if meta.len() != expected_size {
-        return Err(Error::RocksDb(format!(
-            "invalid size {} for {}, expected {}",
-            meta.len(),
-            path,
-            expected_size
-        )));
-    }
-
-    let checksum = calc_crc32(path)?;
-    if checksum == expected_checksum {
-        return Ok(());
-    }
-
-    // RocksDB may have modified the global seqno.
-    let cf_handle = get_cf_handle(db, cf)?;
-    set_external_sst_file_global_seq_no(db, cf_handle, path, 0)?;
-    f.sync_all()
-        .map_err(|e| format!("sync {}: {:?}", path, e))?;
-
-    let checksum = calc_crc32(path)?;
-    if checksum != expected_checksum {
-        return Err(Error::RocksDb(format!(
-            "invalid checksum {} for {}, expected {}",
-            checksum, path, expected_checksum
-        )));
-    }
 
     Ok(())
 }
