+++ src/raftstore/store/snap.rs
@@ -13,10 +13,11 @@ use std::time::Instant;
 use std::{error, result, str, thread, time, u64};
 
 use crc::crc32::{self, Digest, Hasher32};
-use engine::rocks::util::{prepare_sst_for_ingestion, validate_sst_for_ingestion};
 use engine::rocks::Snapshot as DbSnapshot;
 use engine::rocks::DB;
 use engine::{CfName, CF_DEFAULT, CF_LOCK, CF_WRITE};
+use engine_traits::{CFHandleExt, ImportExt};
+use engine_rocks::RocksEngine;
 use kvproto::metapb::Region;
 use kvproto::raft_serverpb::RaftSnapshotData;
 use kvproto::raft_serverpb::{SnapshotCfFile, SnapshotMeta};
@@ -557,10 +558,12 @@ impl Snap {
             if plain_file_used(cf_file.cf) {
                 check_file_size_and_checksum(&cf_file.path, cf_file.size, cf_file.checksum)?;
             } else {
-                prepare_sst_for_ingestion(&cf_file.path, &cf_file.clone_path)?;
-                validate_sst_for_ingestion(
-                    &kv_engine,
-                    cf_file.cf,
+                let engine = RocksEngine::from_ref(&kv_engine);
+                let cf = engine.cf_handle(cf_file.cf)
+                    .ok_or_else(|| Error::Other(box_err!("bad cf handle")))?;
+                engine.prepare_sst_for_ingestion(&cf_file.path, &cf_file.clone_path)?;
+                engine.validate_sst_for_ingestion(
+                    &cf,
                     &cf_file.clone_path,
                     cf_file.size,
                     cf_file.checksum,
